removal of extra functions and variables
functions and variables corresponding to relaxations, power flow problem and tnep problem are removed from the current version. build_mc_data function is also moved from problem file to data.jl.
