+++ Project.toml
@@ -6,9 +6,11 @@ version = "0.0.0"
 
 [deps]
 InfrastructureModels = "2030c09a-7f63-5d83-885d-db604e0e9cc0"
+Ipopt = "b6b21f68-93f8-5de0-b562-5493be1d77c9"
 JuMP = "4076af6c-e467-56ae-b986-b466b2749572"
 Memento = "f28f55f0-a522-5efc-85c2-fe41dfb9b2d9"
 PowerModels = "c36e90e8-916a-50a6-bd94-075b64ef4655"
+PowerModelsACDC = "ff45984e-d068-5f4c-9e32-c4133509d236"
 
 [compat]
 InfrastructureModels = "0.7"
