+++ src/PowerModelsMCDC.jl
@@ -26,7 +26,6 @@ include("core/data.jl")
 include("core/base.jl")
 include("core/constraint.jl")
 include("core/objective.jl")
-include("core/relaxation_scheme.jl")
 include("core/new_multi_conductor_functions.jl")
 include("core/constraint_template.jl")
 include("core/variable_mcdcgrid.jl")
@@ -39,8 +38,6 @@ include("formdcgrid/acp.jl")
 include("formconv/acp.jl")
 
 include("prob/mcdcopf.jl")
-include("prob/mcdcpf.jl")
-
 
 ## Exports
