+++ src/core/objective.jl
@@ -103,34 +103,3 @@ function objective_min_pwl_fuel_cost(pm::_PM.AbstractPowerModel)
         for (n, nw_ref) in _PM.nws(pm))
         )
 end
-
-##################### TNEP Objective   ###################
-function objective_min_cost(pm::_PM.AbstractPowerModel)
-        gen_cost = Dict()
-        for (n, nw_ref) in _PM.nws(pm)
-            for (i,gen) in nw_ref[:gen]
-                pg = _PM.var(pm, n, :pg, i)
-
-                if length(gen["cost"]) == 1
-                    gen_cost[(n,i)] = gen["cost"][1]
-                elseif length(gen["cost"]) == 2
-                    gen_cost[(n,i)] = gen["cost"][1]*pg + gen["cost"][2]
-                elseif length(gen["cost"]) == 3
-                    gen_cost[(n,i)] = gen["cost"][2]*pg + gen["cost"][3]
-                else
-                    gen_cost[(n,i)] = 0.0
-                end
-            end
-        end
-
-        return JuMP.@objective(pm.model, Min,
-            sum(
-                sum(conv["cost"]*_PM.var(pm, n, :conv_ne, i) for (i,conv) in nw_ref[:convdc_ne])
-                +
-                sum(branch["cost"]*_PM.var(pm, n, :branch_ne, i) for (i,branch) in nw_ref[:branchdc_ne])
-                +
-                sum( gen_cost[(n,i)] for (i,gen) in nw_ref[:gen] )
-                for (n, nw_ref) in _PM.nws(pm)
-                    )
-        )
-end
