+++ src/core/variableconv_mc.jl
@@ -420,11 +420,7 @@ function variable_conv_reactor_current_sqr(pm::_PM.AbstractWModels; nw::Int=_PM.
     start = comp_start_value(_PM.ref(pm, nw, :convdc, i), "P_g", c, 1.0)^2
     ) for i in _PM.ids(pm, nw, :convdc)
     )
-    # pprfr = _PM.var(pm, nw)[:pconv_pr_fr] = Dict(i =>JuMP.@variable(pm.model,
-    # [c in 1:_PM.ref(pm, nw, :convdc)[i]["conductors"]], base_name="$(nw)_pconv_pr_fr_$(i)",
-    # start = comp_start_value(_PM.ref(pm, nw, :convdc, i), "P_g", c, 1.0)
-    # ) for i in _PM.ids(pm, nw, :convdc)
-    # )
+
 
     if bounded
         for (c, convdc) in _PM.ref(pm, nw, :convdc)
@@ -561,11 +557,6 @@ function variable_converter_filter_voltage_magnitude_sqr(pm::_PM.AbstractWModels
     start = _PM.ref(pm, nw, :convdc, i, "Vtar")^2
     ) for i in _PM.ids(pm, nw, :convdc)
     )
-    # pprfr = _PM.var(pm, nw)[:pconv_pr_fr] = Dict(i =>JuMP.@variable(pm.model,
-    # [c in 1:_PM.ref(pm, nw, :convdc)[i]["conductors"]], base_name="$(nw)_pconv_pr_fr_$(i)",
-    # start = comp_start_value(_PM.ref(pm, nw, :convdc, i), "P_g", c, 1.0)
-    # ) for i in _PM.ids(pm, nw, :convdc)
-    # )
 
     if bounded
         for (c, convdc) in _PM.ref(pm, nw, :convdc)
@@ -585,11 +576,6 @@ function variable_converter_internal_voltage_cross_products(pm::_PM.AbstractWMod
     start = _PM.ref(pm, nw, :convdc, i, "Vtar")^2
     ) for i in _PM.ids(pm, nw, :convdc)
     )
-    # pprfr = _PM.var(pm, nw)[:pconv_pr_fr] = Dict(i =>JuMP.@variable(pm.model,
-    # [c in 1:_PM.ref(pm, nw, :convdc)[i]["conductors"]], base_name="$(nw)_pconv_pr_fr_$(i)",
-    # start = comp_start_value(_PM.ref(pm, nw, :convdc, i), "P_g", c, 1.0)
-    # ) for i in _PM.ids(pm, nw, :convdc)
-    # )
 
     wicac = _PM.var(pm, nw)[:wic_ac] = Dict(i =>JuMP.@variable(pm.model,
     [c in 1:_PM.ref(pm, nw, :convdc)[i]["conductors"]], base_name="$(nw)_wic_ac_$(i)",
@@ -616,11 +602,7 @@ function variable_converter_internal_voltage_magnitude_sqr(pm::_PM.AbstractWMode
     start = _PM.ref(pm, nw, :convdc, i, "Vtar")^2
     ) for i in _PM.ids(pm, nw, :convdc)
     )
-    # pprfr = _PM.var(pm, nw)[:pconv_pr_fr] = Dict(i =>JuMP.@variable(pm.model,
-    # [c in 1:_PM.ref(pm, nw, :convdc)[i]["conductors"]], base_name="$(nw)_pconv_pr_fr_$(i)",
-    # start = comp_start_value(_PM.ref(pm, nw, :convdc, i), "P_g", c, 1.0)
-    # ) for i in _PM.ids(pm, nw, :convdc)
-    # )
+
     if bounded
         for (c, convdc) in _PM.ref(pm, nw, :convdc)
             JuMP.set_lower_bound.(wcac[c], (convdc["Vmmin"])^2)
@@ -631,472 +613,5 @@ function variable_converter_internal_voltage_magnitude_sqr(pm::_PM.AbstractWMode
 end
 
 function variable_cos_voltage(pm::_PM.AbstractLPACModel; nw::Int=_PM.nw_id_default, bounded::Bool = true, report::Bool=true)
-    #only for lpac
-end
-
-
-
-##################################### TNEP variables ##################################
-"All converter variables"
-function variable_mcdc_converter_ne(pm::_PM.AbstractPowerModel; kwargs...)
-    variable_conv_tranformer_flow_ne(pm; kwargs...)
-    variable_conv_reactor_flow_ne(pm; kwargs...)
-    variable_converter_ne(pm; kwargs...)
-
-    variable_converter_active_power_ne(pm; kwargs...)
-    variable_converter_reactive_power_ne(pm; kwargs...)
-    variable_acside_current_ne(pm; kwargs...)
-    variable_dcside_power_ne(pm; kwargs...)
-    # variable_converter_firing_angle_ne(pm; kwargs...)
-
-    variable_converter_filter_voltage_ne(pm; kwargs...)
-    variable_converter_internal_voltage_ne(pm; kwargs...)
-    #
-    variable_converter_to_grid_active_power_ne(pm; kwargs...)
-    variable_converter_to_grid_reactive_power_ne(pm; kwargs...)
-end
-
-function variable_conv_tranformer_flow_ne(pm::_PM.AbstractPowerModel; kwargs...)
-    variable_conv_transformer_active_power_to_ne(pm; kwargs...)
-    variable_conv_transformer_reactive_power_to_ne(pm; kwargs...)
-end
-
-function variable_conv_reactor_flow_ne(pm::_PM.AbstractPowerModel; kwargs...)
-    variable_conv_reactor_active_power_from_ne(pm; kwargs...)
-    variable_conv_reactor_reactive_power_from_ne(pm; kwargs...)
-end
-
-
-"variable: `0 <= convdc_ne[c] <= 1` for `c` in `candidate converters"
-function variable_converter_ne(pm::_PM.AbstractPowerModel; nw::Int=_PM.nw_id_default, relax::Bool=false, report::Bool=true)
-    if !relax
-        Z_dc_conv_ne = _PM.var(pm, nw)[:conv_ne] = JuMP.@variable(pm.model,
-        [i in _PM.ids(pm, nw, :convdc_ne)], base_name="$(nw)_conv_ne",
-        binary = true,
-        start = _PM.comp_start_value(_PM.ref(pm, nw, :convdc_ne, i), "branchdc_tnep_start", 1.0)
-        )
-    else
-        Z_dc_conv_ne = _PM.var(pm, nw)[:conv_ne] = JuMP.@variable(pm.model,
-        [i in _PM.ids(pm, nw, :convdc_ne)], base_name="$(nw)_conv_ne",
-        lower_bound = 0,
-        upper_bound = 1,
-        start = _PM.comp_start_value(_PM.ref(pm, nw, :convdc_ne, i), "branchdc_tnep_start", 1.0)
-        )
-    end
-    report && _PM.sol_component_value(pm, nw, :convdc_ne, :isbuilt, _PM.ids(pm, nw, :convdc_ne), Z_dc_conv_ne)
- end
-
-"variable: `pconv_ac_ne[j]` for `j` in `candidate convdc`"
-function variable_converter_active_power_ne(pm::_PM.AbstractPowerModel; nw::Int=_PM.nw_id_default, bounded::Bool = true, report::Bool=true)
-    pc_ne = _PM.var(pm, nw)[:pconv_ac_ne] = JuMP.@variable(pm.model,
-    [i in _PM.ids(pm, nw, :convdc_ne)], base_name="$(nw)_pconv_ac_ne",
-    start = _PM.comp_start_value(_PM.ref(pm, nw, :convdc_ne, i), "P_g", 1.0)
-    )
-    if bounded
-        for (c, convdc) in _PM.ref(pm, nw, :convdc_ne)
-            JuMP.set_lower_bound(pc_ne[c],  convdc["Pacmin"])
-            JuMP.set_upper_bound(pc_ne[c],  convdc["Pacmax"])
-        end
-    end
-    report && _PM.sol_component_value(pm, nw, :convdc_ne, :pconv, _PM.ids(pm, nw, :convdc_ne), pc_ne)
- end
-
-"variable: `qconv_ac_ne[j]` for `j` in `candidate convdc`"
-function variable_converter_reactive_power_ne(pm::_PM.AbstractPowerModel; nw::Int=_PM.nw_id_default, bounded::Bool = true, report::Bool=true)
-    qc_ne =  _PM.var(pm, nw)[:qconv_ac_ne] = JuMP.@variable(pm.model,
-    [i in _PM.ids(pm, nw, :convdc_ne)], base_name="$(nw)_qconv_ac_ne",
-    start = _PM.comp_start_value(_PM.ref(pm, nw, :convdc_ne, i), "Q_g", 1.0)
-    )
-    if bounded
-        for (c, convdc) in _PM.ref(pm, nw, :convdc_ne)
-            JuMP.set_lower_bound(qc_ne[c],  convdc["Qacmin"])
-            JuMP.set_upper_bound(qc_ne[c],  convdc["Qacmax"])
-        end
-    end
-    report && _PM.sol_component_value(pm, nw, :convdc_ne, :qconv, _PM.ids(pm, nw, :convdc_ne), qc_ne)
-end
-
-
-"variable: `pconv_grid_ac_to_ne[j]` for `j` in `candidate convdc`"
-function variable_conv_transformer_active_power_to_ne(pm::_PM.AbstractPowerModel; nw::Int=_PM.nw_id_default, bounded::Bool = true, report::Bool=true)
-    bigM = 2;
-    ptfto_ne = _PM.var(pm, nw)[:pconv_tf_to_ne] = JuMP.@variable(pm.model,
-    [i in _PM.ids(pm, nw, :convdc_ne)], base_name="$(nw)_pconv_tf_to_ne",
-    start = _PM.comp_start_value(_PM.ref(pm, nw, :convdc_ne, i), "P_g", 1.0)
-    )
-    if bounded
-        for (c, convdc) in _PM.ref(pm, nw, :convdc_ne)
-            JuMP.set_lower_bound(ptfto_ne[c],  -convdc["Pacrated"] * bigM)
-            JuMP.set_upper_bound(ptfto_ne[c],   convdc["Pacrated"] * bigM)
-        end
-    end
-    report && _PM.sol_component_value(pm, nw, :convdc_ne, :ptf_to, _PM.ids(pm, nw, :convdc_ne), ptfto_ne)
- end
-
-"variable: `qconv_grid_ac_to_ne[j]` for `j` in `candidate convdc`"
-function variable_conv_transformer_reactive_power_to_ne(pm::_PM.AbstractPowerModel; nw::Int=_PM.nw_id_default, bounded::Bool = true, report::Bool=true)
-    bigM = 2;
-    qtfto_ne = _PM.var(pm, nw)[:qconv_tf_to_ne] = JuMP.@variable(pm.model,
-    [i in _PM.ids(pm, nw, :convdc_ne)], base_name="$(nw)_qconv_tf_to_ne",
-    start = _PM.comp_start_value(_PM.ref(pm, nw, :convdc_ne, i), "Q_g", 1.0)
-    )
-    if bounded
-        for (c, convdc) in _PM.ref(pm, nw, :convdc_ne)
-            JuMP.set_lower_bound(qtfto_ne[c],  -convdc["Qacrated"] * bigM)
-            JuMP.set_upper_bound(qtfto_ne[c],   convdc["Qacrated"] * bigM)
-        end
-    end
-    report && _PM.sol_component_value(pm, nw, :convdc_ne, :qtf_to, _PM.ids(pm, nw, :convdc_ne), qtfto_ne)
-end
-
-
-"variable: `pconv_pr_from_ne[j]` for `j` in `candidate convdc`"
-function variable_conv_reactor_active_power_from_ne(pm::_PM.AbstractPowerModel; nw::Int=_PM.nw_id_default, bounded::Bool = true, report::Bool=true)
-    bigM = 2
-    pprfr_ne = _PM.var(pm, nw)[:pconv_pr_fr_ne] = JuMP.@variable(pm.model,
-    [i in _PM.ids(pm, nw, :convdc_ne)], base_name="$(nw)_pconv_pr_from_ne",
-    start = _PM.comp_start_value(_PM.ref(pm, nw, :convdc_ne, i), "P_g", 1.0)
-    )
-    if bounded
-        for (c, convdc) in _PM.ref(pm, nw, :convdc_ne)
-            JuMP.set_lower_bound(pprfr_ne[c],  -convdc["Pacrated"] * bigM)
-            JuMP.set_upper_bound(pprfr_ne[c],   convdc["Pacrated"] * bigM)
-        end
-    end
-    report && _PM.sol_component_value(pm, nw, :convdc_ne, :ppr_fr, _PM.ids(pm, nw, :convdc_ne), pprfr_ne)
-end
-
-"variable: `qconv_pr_from_ne[j]` for `j` in `candidate convdc`"
-function variable_conv_reactor_reactive_power_from_ne(pm::_PM.AbstractPowerModel; nw::Int=_PM.nw_id_default, bounded::Bool = true, report::Bool=true)
-    bigM = 2
-    qprfr_ne = _PM.var(pm, nw)[:qconv_pr_fr_ne] = JuMP.@variable(pm.model,
-    [i in _PM.ids(pm, nw, :convdc_ne)], base_name="$(nw)_qconv_pr_from_ne",
-    start = _PM.comp_start_value(_PM.ref(pm, nw, :convdc_ne, i), "Q_g", 1.0)
-    )
-    if bounded
-       for (c, convdc) in _PM.ref(pm, nw, :convdc_ne)
-           JuMP.set_lower_bound(qprfr_ne[c],  -convdc["Qacrated"] * bigM)
-           JuMP.set_upper_bound(qprfr_ne[c],   convdc["Qacrated"] * bigM)
-       end
-   end
-   report && _PM.sol_component_value(pm, nw, :convdc_ne, :qpr_fr, _PM.ids(pm, nw, :convdc_ne), qprfr_ne)
-end
-
-
-"variable: `pconv_grid_ac_ne[j]` for `j` in `candidate convdc`"
-function variable_converter_to_grid_active_power_ne(pm::_PM.AbstractPowerModel; nw::Int=_PM.nw_id_default, bounded::Bool = true, report::Bool=true)
-    bigM = 2
-    ptffr_ne = _PM.var(pm, nw)[:pconv_tf_fr_ne] = JuMP.@variable(pm.model,
-    [i in _PM.ids(pm, nw, :convdc_ne)], base_name="$(nw)_pconv_tf_fr_ne",
-    start = _PM.comp_start_value(_PM.ref(pm, nw, :convdc_ne, i), "P_g", 1.0)
-    )
-    if bounded
-       for (c, convdc) in _PM.ref(pm, nw, :convdc_ne)
-           JuMP.set_lower_bound(ptffr_ne[c],  -convdc["Pacrated"] * bigM)
-           JuMP.set_upper_bound(ptffr_ne[c],   convdc["Pacrated"] * bigM)
-       end
-   end
-   report && _PM.sol_component_value(pm, nw, :convdc_ne, :pgrid, _PM.ids(pm, nw, :convdc_ne), ptffr_ne)
-end
-
-"variable: `qconv_grid_ac_ne[j]` for `j` in `candidate convdc`"
-function variable_converter_to_grid_reactive_power_ne(pm::_PM.AbstractPowerModel; nw::Int=_PM.nw_id_default, bounded::Bool = true, report::Bool=true)
-    bigM = 2
-    qtffr_ne = _PM.var(pm, nw)[:qconv_tf_fr_ne] = JuMP.@variable(pm.model,
-        [i in _PM.ids(pm, nw, :convdc_ne)], base_name="$(nw)_qconv_tf_fr_ne",
-        start = _PM.comp_start_value(_PM.ref(pm, nw, :convdc_ne, i), "Q_g", 1.0)
-        )
-        if bounded
-           for (c, convdc) in _PM.ref(pm, nw, :convdc_ne)
-               JuMP.set_lower_bound(qtffr_ne[c],  -convdc["Qacrated"] * bigM)
-               JuMP.set_upper_bound(qtffr_ne[c],   convdc["Qacrated"] * bigM)
-           end
-       end
-       report && _PM.sol_component_value(pm, nw, :convdc_ne, :qgrid, _PM.ids(pm, nw, :convdc_ne), qtffr_ne)
-end
-
-
-"variable: `pconv_dc_ne[j]` for `j` in `candidate convdc`"
-function variable_dcside_power_ne(pm::_PM.AbstractPowerModel; nw::Int=_PM.nw_id_default, bounded::Bool = true, report::Bool=true)
-    bigM = 1.2; # to account for losses, maximum losses to be derived
-    pcdc_ne = _PM.var(pm, nw)[:pconv_dc_ne] = JuMP.@variable(pm.model,
-    [i in _PM.ids(pm, nw, :convdc_ne)], base_name="$(nw)_pconv_dc_ne",
-    start = _PM.comp_start_value(_PM.ref(pm, nw, :convdc_ne, i), "Pdcset", 1.0)
-    )
-    if bounded
-        for (c, convdc) in _PM.ref(pm, nw, :convdc_ne)
-            JuMP.set_lower_bound(pcdc_ne[c],  -convdc["Pacrated"] * bigM)
-            JuMP.set_upper_bound(pcdc_ne[c],   convdc["Pacrated"] * bigM)
-        end
-    end
-    report && _PM.sol_component_value(pm, nw, :convdc_ne, :pdc, _PM.ids(pm, nw, :convdc_ne), pcdc_ne)
- end
-
-"variable: `pconv_dc_ne[j]` for `j` in `candidate convdc`"
-function variable_converter_firing_angle_ne(pm::_PM.AbstractPowerModel; nw::Int=_PM.nw_id_default, bounded::Bool = true, report::Bool=true)
-    phic_ne = _PM.var(pm, nw)[:phiconv_ne] = JuMP.@variable(pm.model,
-    [i in _PM.ids(pm, nw, :convdc_ne)], base_name="$(nw)_phiconv",
-    lower_bound = 0,
-    upper_bound =  pi,
-    start = acos(_PM.comp_start_value(_PM.ref(pm, nw, :convdc_ne, i), "Pdcset", 1.0) / sqrt((_PM.comp_start_value(_PM.ref(pm, nw, :convdc_ne, i), "Pacrated", 1.0))^2 + (_PM.comp_start_value(_PM.ref(pm, nw, :convdc_ne, i), "Qacrated", 1.0))^2))
-    )
-    if bounded
-        for (c, convdc) in _PM.ref(pm, nw, :convdc_ne)
-            JuMP.set_lower_bound(phic_ne[c],  0)
-            JuMP.set_upper_bound(phic_ne[c],  pi)
-        end
-    end
-    report && _PM.sol_component_value(pm, nw, :convdc_ne, :phi, _PM.ids(pm, nw, :convdc_ne), phic_ne)
-end
-
-"variable: `iconv_ac_ne[j]` for `j` in `candidate convdc`"
-function variable_acside_current_ne(pm::_PM.AbstractPowerModel; nw::Int=_PM.nw_id_default, bounded::Bool = true, report::Bool=true)
-    ic_ne = _PM.var(pm, nw)[:iconv_ac_ne] = JuMP.@variable(pm.model,
-    [i in _PM.ids(pm, nw, :convdc_ne)], base_name="$(nw)_iconv_ac_ne",
-    start = _PM.comp_start_value(_PM.ref(pm, nw, :convdc_ne, i), "P_g", 1.0)
-    )
-    if bounded
-        for (c, convdc) in _PM.ref(pm, nw, :convdc_ne)
-            JuMP.set_lower_bound(ic_ne[c],  0)
-            JuMP.set_upper_bound(ic_ne[c],  convdc["Imax"])
-        end
-    end
-
-    report && _PM.sol_component_value(pm, nw, :convdc_ne, :iconv, _PM.ids(pm, nw, :convdc_ne), ic_ne)
- end
-
-"variable: `iconv_ac_ne[j]` and `iconv_ac_sq_ne[j]` for `j` in `candidate convdc`"
-function variable_acside_current_ne(pm::_PM.AbstractWModels; nw::Int=_PM.nw_id_default, bounded::Bool = true, report::Bool=true)
-    ic_ne = _PM.var(pm, nw)[:iconv_ac_ne] = JuMP.@variable(pm.model,
-    [i in _PM.ids(pm, nw, :convdc_ne)], base_name="$(nw)_iconv_ac_ne",
-    start = _PM.comp_start_value(_PM.ref(pm, nw, :convdc_ne, i), "P_g", 1.0)
-    )
-    icsq_ne = _PM.var(pm, nw)[:iconv_ac_sq_ne] = JuMP.@variable(pm.model,
-    [i in _PM.ids(pm, nw, :convdc_ne)], base_name="$(nw)_iconv_ac_sq_ne",
-    start = _PM.comp_start_value(_PM.ref(pm, nw, :convdc_ne, i), "P_g", 1.0)
-    )
-    if bounded
-    for (c, convdc) in _PM.ref(pm, nw, :convdc_ne)
-        JuMP.set_lower_bound(ic_ne[c],  0)
-        JuMP.set_upper_bound(ic_ne[c],  convdc["Imax"])
-        JuMP.set_lower_bound(icsq_ne[c],  0)
-        JuMP.set_upper_bound(icsq_ne[c],  convdc["Imax"]^2)
-    end
-end
-report && _PM.sol_component_value(pm, nw, :convdc_ne, :iconv_ac, _PM.ids(pm, nw, :convdc_ne), ic_ne)
-report && _PM.sol_component_value(pm, nw, :convdc_ne, :iconv_ac_sq, _PM.ids(pm, nw, :convdc_ne), icsq_ne)
-end
-
-"variable: `itf_sq_ne[j]` for `j` in `candidate convdc`"
-function variable_conv_transformer_current_sqr_ne(pm::_PM.AbstractWModels;  nw::Int=_PM.nw_id_default, bounded::Bool = true, report::Bool=true)
-    bigM = 2 #TODO derive exact bound
-    itfsq_ne = _PM.var(pm, nw)[:itf_sq_ne] = JuMP.@variable(pm.model,
-    [i in _PM.ids(pm, nw, :convdc_ne)], base_name="$(nw)_itf_sq_ne",
-    start = _PM.comp_start_value(_PM.ref(pm, nw, :convdc_ne, i), "P_g", 1.0)^2
-    )
-    if bounded
-        for (c, convdc) in _PM.ref(pm, nw, :convdc_ne)
-            JuMP.set_lower_bound(itfsq_ne[c],  0)
-            JuMP.set_upper_bound(itfsq_ne[c], (bigM * convdc["Imax"])^2)
-        end
-    end
-report && _PM.sol_component_value(pm, nw, :convdc_ne, :itf_sq, _PM.ids(pm, nw, :convdc_ne), itfsq_ne)
-end
-
-
-"variable: `irc_sq_ne[j]` for `j` in `candidate convdc`"
-function variable_conv_reactor_current_sqr_ne(pm::_PM.AbstractWModels;  nw::Int=_PM.nw_id_default, bounded::Bool = true, report::Bool=true)
-    bigM = 2 #TODO derive exact bound
-    iprsq_ne = _PM.var(pm, nw)[:irc_sq_ne] = JuMP.@variable(pm.model,
-    [i in _PM.ids(pm, nw, :convdc_ne)], base_name="$(nw)_irc_sq_ne",
-    start = _PM.comp_start_value(_PM.ref(pm, nw, :convdc_ne, i), "P_g", 1.0)^2
-    )
-    if bounded
-        for (c, convdc) in _PM.ref(pm, nw, :convdc_ne)
-            JuMP.set_lower_bound(iprsq_ne[c],  0)
-            JuMP.set_upper_bound(iprsq_ne[c], (bigM * convdc["Imax"])^2)
-        end
-    end
-report && _PM.sol_component_value(pm, nw, :convdc_ne, :ipr_sq, _PM.ids(pm, nw, :convdc_ne), iprsq_ne)
-end
-
-
-function variable_converter_filter_voltage_ne(pm::_PM.AbstractPowerModel; kwargs...)
-    variable_converter_filter_voltage_magnitude_ne(pm; kwargs...)
-    variable_converter_filter_voltage_angle_ne(pm; kwargs...)
-end
-
-
-"variable: `vmf_ne[j]` for `j` in `candidate convdc`"
-function variable_converter_filter_voltage_magnitude_ne(pm::_PM.AbstractPowerModel; nw::Int=_PM.nw_id_default, bounded::Bool = true, report::Bool=true)
-    bigM = 1.2; # only internal converter voltage is strictly regulated
-    vmf_ne = _PM.var(pm, nw)[:vmf_ne] = JuMP.@variable(pm.model,
-    [i in _PM.ids(pm, nw, :convdc_ne)], base_name="$(nw)_vmf_ne",
-    start = _PM.ref(pm, nw, :convdc_ne, i, "Vtar")
-    )
-    if bounded
-        for (c, convdc) in _PM.ref(pm, nw, :convdc_ne)
-            JuMP.set_lower_bound(vmf_ne[c], convdc["Vmmin"] / bigM)
-            JuMP.set_upper_bound(vmf_ne[c], convdc["Vmmax"] * bigM)
-        end
-    end
-    report && _PM.sol_component_value(pm, nw, :convdc_ne, :vmfilt, _PM.ids(pm, nw, :convdc_ne), vmf_ne)
-end
-
-
-"variable: `vaf_ne[j]` for `j` in `candidate convdc`"
-function variable_converter_filter_voltage_angle_ne(pm::_PM.AbstractPowerModel; nw::Int=_PM.nw_id_default, bounded::Bool = true, report::Bool=true)
-    bigM = 2*pi; #
-    vaf_ne = _PM.var(pm, nw)[:vaf_ne] = JuMP.@variable(pm.model,
-    [i in _PM.ids(pm, nw, :convdc_ne)], base_name="$(nw)_vaf_ne",
-    start = 0
-    )
-    if bounded
-        for (c, convdc) in _PM.ref(pm, nw, :convdc_ne)
-            JuMP.set_lower_bound(vaf_ne[c], -bigM)
-            JuMP.set_upper_bound(vaf_ne[c],  bigM)
-        end
-    end
-    report && _PM.sol_component_value(pm, nw, :convdc_ne, :vafilt, _PM.ids(pm, nw, :convdc_ne), vaf_ne)
-end
-
-
-function variable_converter_internal_voltage_ne(pm::_PM.AbstractPowerModel; kwargs...)
-    variable_converter_internal_voltage_magnitude_ne(pm; kwargs...)
-    variable_converter_internal_voltage_angle_ne(pm; kwargs...)
-end
-
-
-"variable: `vmc_ne[j]` for `j` in `candidate convdc`"
-function variable_converter_internal_voltage_magnitude_ne(pm::_PM.AbstractPowerModel; nw::Int=_PM.nw_id_default, bounded::Bool = true, report::Bool=true)
-    vmc_ne = _PM.var(pm, nw)[:vmc_ne] = JuMP.@variable(pm.model,
-    [i in _PM.ids(pm, nw, :convdc_ne)], base_name="$(nw)_vmc_ne",
-    start = _PM.ref(pm, nw, :convdc_ne, i, "Vtar")
-    )
-    if bounded
-        for (c, convdc) in _PM.ref(pm, nw, :convdc)
-            JuMP.set_lower_bound(vmc_ne[c], convdc["Vmmin"])
-            JuMP.set_upper_bound(vmc_ne[c], convdc["Vmmax"])
-        end
-    end
-    report && _PM.sol_component_value(pm, nw, :convdc_ne, :vmconv, _PM.ids(pm, nw, :convdc_ne), vmc_ne)
-end
-
-"variable: `vac_ne[j]` for `j` in `candidate convdc`"
-function variable_converter_internal_voltage_angle_ne(pm::_PM.AbstractPowerModel; nw::Int=_PM.nw_id_default, bounded::Bool = true, report::Bool=true)
-    bigM = 2*pi; #
-    vac_ne = _PM.var(pm, nw)[:vac_ne] = JuMP.@variable(pm.model,
-    [i in _PM.ids(pm, nw, :convdc_ne)], base_name="$(nw)_vac_ne",
-    start = 0
-    )
-    if bounded
-        for (c, convdc) in _PM.ref(pm, nw, :convdc_ne)
-            JuMP.set_lower_bound(vac_ne[c], -bigM)
-            JuMP.set_upper_bound(vac_ne[c],  bigM)
-        end
-    end
-    report && _PM.sol_component_value(pm, nw, :convdc_ne, :vaconv, _PM.ids(pm, nw, :convdc_ne), vac_ne)
-end
-
-
-
-"variable: `wrf_ac_ne[j]` and `wif_ac`  for `j` in `candidate convdc`"
-function variable_converter_filter_voltage_cross_products_ne(pm::_PM.AbstractPowerModel; nw::Int=_PM.nw_id_default, bounded::Bool = true, report::Bool=true)
-    bigM = 1.2; # only internal converter voltage is strictly regulated
-    wrfac_ne = _PM.var(pm, nw)[:wrf_ac_ne] = JuMP.@variable(pm.model,
-    [i in _PM.ids(pm, nw, :convdc_ne)], base_name="$(nw)_wrf_ac_ne",
-    start = _PM.ref(pm, nw, :convdc_ne, i, "Vtar")^2
-    )
-    wifac_ne = _PM.var(pm, nw)[:wif_ac_ne] = JuMP.@variable(pm.model,
-    [i in _PM.ids(pm, nw, :convdc_ne)], base_name="$(nw)_wif_ac_ne",
-    start = _PM.ref(pm, nw, :convdc_ne, i, "Vtar")^2
-    )
-    if bounded
-        for (c, convdc) in _PM.ref(pm, nw, :convdc_ne)
-            JuMP.set_lower_bound(wrfac_ne[c],  0)
-            JuMP.set_upper_bound(wrfac_ne[c],  (convdc["Vmmax"] * bigM)^2)
-            JuMP.set_lower_bound(wifac_ne[c], -(convdc["Vmmax"] * bigM)^2)
-            JuMP.set_upper_bound(wifac_ne[c],  (convdc["Vmmax"] * bigM)^2)
-        end
-    end
-    report && _PM.sol_component_value(pm, nw, :convdc_ne, :wrfilt, _PM.ids(pm, nw, :convdc_ne), wrfac_ne)
-    report && _PM.sol_component_value(pm, nw, :convdc_ne, :wifilt, _PM.ids(pm, nw, :convdc_ne), wifac_ne)
-end
-
-"variable: `wf_ac_ne` for `j` in `candidate convdc`"
-function variable_converter_filter_voltage_magnitude_sqr_ne(pm::_PM.AbstractPowerModel; nw::Int=_PM.nw_id_default, bounded::Bool = true, report::Bool=true)
-    bigM = 1.2; # only internal converter voltage is strictly regulated
-    wfac_ne = _PM.var(pm, nw)[:wf_ac_ne] = JuMP.@variable(pm.model,
-    [i in _PM.ids(pm, nw, :convdc_ne)], base_name="$(nw)_wf_ac_ne",
-    start = _PM.ref(pm, nw, :convdc_ne, i, "Vtar")^2
-    )
-    if bounded
-        for (c, convdc) in _PM.ref(pm, nw, :convdc_ne)
-            JuMP.set_lower_bound(wfac_ne[c], 0) # check
-            # JuMP.set_lower_bound(wfac_ne[c], (convdc["Vmmin"] / bigM)^2)
-            JuMP.set_upper_bound(wfac_ne[c], (convdc["Vmmax"] * bigM)^2)
-        end
-    end
-    report && _PM.sol_component_value(pm, nw, :convdc_ne, :wfilt, _PM.ids(pm, nw, :convdc_ne), wfac_ne)
-end
-
-
-"variable: `wrc_ac_ne[j]` and `wic_ac_ne[j]`  for `j` in `candidate convdc`"
-function variable_converter_internal_voltage_cross_products_ne(pm::_PM.AbstractPowerModel; nw::Int=_PM.nw_id_default, bounded::Bool = true, report::Bool=true)
-    bigM = 1.2; # only internal converter voltage is strictly regulated
-    wrcac_ne = _PM.var(pm, nw)[:wrc_ac_ne] = JuMP.@variable(pm.model,
-    [i in _PM.ids(pm, nw, :convdc_ne)], base_name="$(nw)_wrc_ac_ne",
-    start = _PM.ref(pm, nw, :convdc_ne, i, "Vtar")^2
-    )
-    wicac_ne = _PM.var(pm, nw)[:wic_ac_ne] = JuMP.@variable(pm.model,
-    [i in _PM.ids(pm, nw, :convdc_ne)], base_name="$(nw)_wic_ac_ne",
-    start = _PM.ref(pm, nw, :convdc_ne, i, "Vtar")^2
-    )
-    if bounded
-        for (c, convdc) in _PM.ref(pm, nw, :convdc_ne)
-            JuMP.set_lower_bound(wrcac_ne[c],  0)
-            JuMP.set_upper_bound(wrcac_ne[c],  (convdc["Vmmax"] * bigM)^2)
-            JuMP.set_lower_bound(wicac_ne[c], -(convdc["Vmmax"] * bigM)^2)
-            JuMP.set_upper_bound(wicac_ne[c],  (convdc["Vmmax"] * bigM)^2)
-        end
-    end
-
-    report && _PM.sol_component_value(pm, nw, :convdc_ne, :wrconv, _PM.ids(pm, nw, :convdc_ne), wrcac_ne)
-    report && _PM.sol_component_value(pm, nw, :convdc_ne, :wiconv, _PM.ids(pm, nw, :convdc_ne), wicac_ne)
-
-end
-
-"variable: `wc_ac_ne[j]` for `j` in `candidate convdc`"
-function variable_converter_internal_voltage_magnitude_sqr_ne(pm::_PM.AbstractPowerModel; nw::Int=_PM.nw_id_default, bounded::Bool = true, report::Bool=true)
-    wcac_ne = _PM.var(pm, nw)[:wc_ac_ne] = JuMP.@variable(pm.model,
-    [i in _PM.ids(pm, nw, :convdc_ne)], base_name="$(nw)_wc_ac_ne",
-    start = _PM.ref(pm, nw, :convdc_ne, i, "Vtar")^2
-    )
-    if bounded
-        for (c, convdc) in _PM.ref(pm, nw, :convdc_ne)
-            JuMP.set_lower_bound(wcac_ne[c], 0) # check
-            # JuMP.set_lower_bound(wcac[c], (convdc["Vmmin"])^2)
-            JuMP.set_upper_bound(wcac_ne[c], (convdc["Vmmax"])^2)
-        end
-    end
-    report && _PM.sol_component_value(pm, nw, :convdc_ne, :wconv, _PM.ids(pm, nw, :convdc_ne), wcac_ne)
-end
-
-function variable_voltage_slack(pm::_PM.AbstractWModels; nw::Int=_PM.nw_id_default, bounded::Bool = true, report::Bool=false)
-    w_du = _PM.var(pm, nw)[:w_du] = JuMP.@variable(pm.model,
-    [i in _PM.ids(pm, nw, :convdc_ne)], base_name="$(nw)_w_du",
-    start = 0,  # check
-    )
-    if bounded
-        for (c, convdc) in _PM.ref(pm, nw, :convdc_ne)
-            JuMP.set_lower_bound(w_du[c], 0) # check
-            JuMP.set_upper_bound(w_du[c], 2) # check
-        end
-    end
-    report && _PM.sol_component_value(pm, nw, :convdc_ne, :wdu, _PM.ids(pm, nw, :convdc_ne), w_du)
-end
-
-
-function variable_cos_voltage_ne(pm::_PM.AbstractPowerModel; nw::Int=_PM.nw_id_default, bounded::Bool = true, report::Bool=true)
     #only for lpac
 end
