+++ src/prob/mcdcopf.jl
@@ -3,7 +3,7 @@ export solve_mcdcopf
 ""
 function solve_mcdcopf(file::String, model_type::Type, solver; kwargs...)
     data = _PM.parse_file(file)
-    process_additional_data!(data)
+    PowerModelsMCDC.build_mc_data!(data)
     return solve_mcdcopf(data, model_type, solver; ref_extensions = [add_ref_dcgrid!], kwargs...)
 end
